Make textfield for name a label and fix bug in GPF
The galaxy/zone name in the properties is now a label, so that it's clear there's no way to edit it. If a way to edit it is added in the future, this can be reverted.

Any time you tried to add/remove layers from a galaxy that has an invalid root name in the properties, it'd silently fail. This addresses that problem AND makes sure any failure while saving the map file isn't silent.